backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
strawberry-graphql[fastapi]
aiofiles
